Renamed variable_private.*pp to better reflect its role
