+++ projectfiles/CodeBlocks/wesnoth.cbp
@@ -380,8 +380,8 @@
 		<Unit filename="../../src/formula/tokenizer.hpp" />
 		<Unit filename="../../src/formula/variant.cpp" />
 		<Unit filename="../../src/formula/variant.hpp" />
-		<Unit filename="../../src/formula/variant_private.cpp" />
-		<Unit filename="../../src/formula/variant_private.hpp" />
+		<Unit filename="../../src/formula/variant_value.cpp" />
+		<Unit filename="../../src/formula/variant_value.hpp" />
 		<Unit filename="../../src/game_board.cpp" />
 		<Unit filename="../../src/game_board.hpp" />
 		<Unit filename="../../src/game_classification.cpp" />
