+++ source_lists/wesnoth
@@ -112,7 +112,7 @@ formula/function.cpp
 formula/string_utils.cpp
 formula/tokenizer.cpp
 formula/variant.cpp
-formula/variant_private.cpp
+formula/variant_value.cpp
 game_board.cpp
 game_classification.cpp
 game_config_manager.cpp
